mesextended: tidyup and more (#56)
* mesextended: tidyup and more

added left click trade and left click follow

cleaned up code

removed left click report

* add disabledby

* bugfix

* Update menuentryswapperextended.gradle.kts
