+++ menuentryswapperextended/src/main/java/net/runelite/client/plugins/menuentryswapperextended/util/teleportmode/BurningAmuletMode.java
@@ -1,5 +1,6 @@
 /*
- * Copyright (c) 2019, Alan Baumgartner <https://github.com/alanbaumgartner>
+ * Copyright (c) 2018, https://openosrs.com
+ * Copyright (c) 2018, Kyle <https://github.com/kyleeld>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -22,19 +23,17 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.menuentryswapperextended.util;
+package net.runelite.client.plugins.menuentryswapperextended.util.teleportmode;
 
-public enum GamesNecklaceMode
+public enum BurningAmuletMode
 {
-	BURTHORPE("Burthorpe"),
-	BARBARIAN_OUTPOST("Barbarian Outpost"),
-	CORPOREAL_BEAST("Corporeal Beast"),
-	TEARS_OF_GUTHIX("Tears of Guthix"),
-	WINTER("Wintertodt Camp");
+	CHAOS_TEMPLE("Chaos Temple"),
+	BANDIT_CAMP("Bandit Camp"),
+	LAVA_MAZE("Lava Maze");
 
 	private final String name;
 
-	GamesNecklaceMode(String name)
+	BurningAmuletMode(String name)
 	{
 		this.name = name;
 	}
