+++ menuentryswapperextended/src/main/java/net/runelite/client/plugins/menuentryswapperextended/util/teleportmode/GamesNecklaceMode.java
@@ -22,17 +22,19 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.menuentryswapperextended.util;
+package net.runelite.client.plugins.menuentryswapperextended.util.teleportmode;
 
-public enum DuelingRingMode
+public enum GamesNecklaceMode
 {
-	DUEL_ARENA("Duel Arena"),
-	CASTLE_WARS("Castle Wars"),
-	CLAN_WARS("Clan Wars");
+	BURTHORPE("Burthorpe"),
+	BARBARIAN_OUTPOST("Barbarian Outpost"),
+	CORPOREAL_BEAST("Corporeal Beast"),
+	TEARS_OF_GUTHIX("Tears of Guthix"),
+	WINTER("Wintertodt Camp");
 
 	private final String name;
 
-	DuelingRingMode(String name)
+	GamesNecklaceMode(String name)
 	{
 		this.name = name;
 	}
